Margin and Padding with Display types
Margin and Padding were set to the three cards. Margin was used to align a button which had a display type
set to block such that it would be on a new line. Margin values were 0 and auto to align the button in the
center. Just specifying auto will not align it in the center since margin takes into account height and width
and needs a value of zero for height to align it horizontally, such that width takes the value auto